refactor(web-app): streamline imports across components and hooks
- Consolidated imports from individual files to index files for better organization.
- Updated references in various components, hooks, and providers to reflect the new import structure.
- Improved code readability and maintainability by reducing import paths.
