+++ apps/web-app/src/components/ui/index.ts
@@ -1,0 +1,2 @@
+export * from "./WalletButton";
+export * from "./hover-border-gradient";
