+++ apps/web-app/src/features/pools/components/pages/PoolDetail.tsx
@@ -6,7 +6,7 @@ import Link from "next/link";
 import { orchestrator, usePoolInfo, useUserPosition } from "@/lib/orchestrator";
 import type { PoolAction } from "@/lib/orchestrator";
 import { fromSmallestUnit } from "@/lib/helpers/tokenUtils";
-import { useWallet } from "@/hooks/useWallet";
+import { useWallet } from "@/hooks";
 import { PoolActionModal } from "../PoolActionModal";
 
 interface PoolDetailProps {
