+++ apps/web-app/src/features/swap/components/pages/Swap.tsx
@@ -1,18 +1,20 @@
 "use client";
 
 import React, { useEffect, useMemo } from "react";
-import { useWallet } from "@/hooks/useWallet";
-import { useTokenBalance } from "@/hooks/useTokenBalance";
-import { useEVMTokenBalance } from "@/hooks/useEVMTokenBalance";
-import { useGasCheck } from "@/hooks/useGasCheck";
-import { useWalletType } from "@/hooks/useWalletType";
+import {
+  useWallet,
+  useTokenBalance,
+  useEVMTokenBalance,
+  useGasCheck,
+  useWalletType,
+} from "@/hooks";
 import { useWalletClient, useChainId } from "wagmi";
 import {
   getTokens,
   getAvailableTokens,
   type Token,
 } from "@/lib/helpers/stellar/soroswap";
-import type { EVMToken } from "@/lib/types/evmToken";
+import type { EVMToken } from "@/lib/types";
 import { SUPPORTED_CHAINS } from "@/lib/constants/evmConfig";
 import { extractContractErrorOrNull } from "@/lib/helpers/stellar/contractErrors";
 import { OrderManagement } from "../ui/OrderManagement";
