+++ apps/web-app/src/lib/types/index.ts
@@ -1,0 +1,6 @@
+export * from "./cowswapTypes";
+export * from "./evmToken";
+export * from "./lendingTypes";
+export * from "./priceTypes";
+export * from "./soroswapTypes";
+export * from "./tokenTypes";
